Document, review and cut line lenghts in all actions I created
