+++ applications/conceptual-model-editor/src/action/add-class-to-visual-model.ts
@@ -1,15 +1,14 @@
 import { WritableVisualModel } from "@dataspecer/core-v2/visual-model";
 import { isSemanticModelClass } from "@dataspecer/core-v2/semantic-model/concepts";
 
-import { withAggregatedEntity } from "./utilities";
+import { getVisualNodeContentBasedOnExistingEntities, withAggregatedEntity } from "./utilities";
 import { UseNotificationServiceWriterType } from "../notification/notification-service-context";
 import { ModelGraphContextType } from "../context/model-context";
 import { UseDiagramType } from "../diagram/diagram-hook";
 import { addRelatedEntitiesAction } from "./add-related-entities-to-visual-model";
-import { findPositionForNewNodesUsingLayouting } from "./layout-visual-model";
+import { findPositionForNewNodesUsingLayouting, findPositionForNewNodeUsingLayouting } from "./layout-visual-model";
 import { ClassesContextType } from "../context/classes-context";
 import { addVisualNode } from "../dataspecer/visual-model/operation/add-visual-node";
-import { getVisualNodeContentBasedOnExistingEntities } from "./add-semantic-attribute-to-visual-model";
 
 export async function addSemanticClassToVisualModelAction(
   notifications: UseNotificationServiceWriterType,
@@ -23,13 +22,13 @@ export async function addSemanticClassToVisualModelAction(
 ) {
   const entities = graph.aggregatorView.getEntities();
   if (position === null) {
-    position = (await findPositionForNewNodesUsingLayouting(notifications, diagram, graph, visualModel, classes, [entityIdentifier]))[entityIdentifier];
+    position = await findPositionForNewNodeUsingLayouting(
+      notifications, diagram, graph, visualModel, classes, entityIdentifier);
   }
 
   withAggregatedEntity(notifications, entities,
     entityIdentifier, modelIdentifier,
     isSemanticModelClass, (entity) => {
-      // TODO PRQuestion: How to handle this? Put it into the addVisualNode?
       const content = getVisualNodeContentBasedOnExistingEntities(
         classes, entity);
       addVisualNode(
