+++ applications/conceptual-model-editor/src/action/add-semantic-attribute-to-visual-node.ts
@@ -1,7 +1,12 @@
 import { VisualNode, WritableVisualModel } from "@dataspecer/core-v2/visual-model";
 import { UseNotificationServiceWriterType } from "../notification/notification-service-context";
 
-// TODO RadStr: Document
+/**
+ * Adds semantic {@link attribute} to {@link domainNode} on given {@link position}.
+ * If the attribute is already present, it is not added and error is either reported or not
+ * based on the {@link shouldReportDuplicateAsNotification}.
+ * @param position is the position to put the attribute on, if null it is put at the end.
+ */
 export function addSemanticAttributeToVisualNodeAction(
   notifications: UseNotificationServiceWriterType,
   visualModel: WritableVisualModel,
@@ -9,7 +14,7 @@ export function addSemanticAttributeToVisualNodeAction(
   attribute: string,
   position: number | null,
   shouldReportDuplicateAsNotification: boolean,
-) {
+): void {
   if(domainNode.content.includes(attribute)) {
     if(shouldReportDuplicateAsNotification) {
       notifications.error("The given attribute to be shown is already present on the visual node");
