+++ applications/conceptual-model-editor/src/action/change-visual-model.ts
@@ -1,6 +1,9 @@
 import { QueryParamsContextType } from "../context/query-params-context";
 import { ModelGraphContextType } from "../context/model-context";
 
+/**
+ * Changes visual model to the {@link viewIdentifier} and updates url.
+ */
 export function changeVisualModelAction (
   graph: ModelGraphContextType,
   queryParamsContext: QueryParamsContextType,
