+++ applications/conceptual-model-editor/src/action/create-new-visual-model-from-source-visual-model.ts
@@ -1,4 +1,12 @@
-import { isVisualNode, isVisualProfileRelationship, isVisualRelationship, VisualModel, VisualProfileRelationship, VisualRelationship, WritableVisualModel } from "@dataspecer/core-v2/visual-model";
+import {
+  isVisualNode,
+  isVisualProfileRelationship,
+  isVisualRelationship,
+  VisualModel,
+  VisualProfileRelationship,
+  VisualRelationship,
+  WritableVisualModel
+} from "@dataspecer/core-v2/visual-model";
 import { ModelGraphContextType, UseModelGraphContextType } from "../context/model-context";
 import { UseNotificationServiceWriterType } from "../notification/notification-service-context";
 import { LanguageString } from "@dataspecer/core/core/core-resource";
@@ -96,7 +104,8 @@ function addEdgesFromSourceModelToTargetModel(
 /**
  * Creates new relationship based on given mapping
  */
-function createRelationshipCopy<T extends Omit<VisualRelationship, "identifier" | "type"> | Omit<VisualProfileRelationship, "identifier" | "type">>(
+function createRelationshipCopy<T extends Omit<VisualRelationship, "identifier" | "type"> |
+                                         Omit<VisualProfileRelationship, "identifier" | "type">>(
   oldVisualRelationship: T,
   oldNodeIdToNewNodeIdMapping: Record<string, string>,
 ): T | null {
