+++ applications/conceptual-model-editor/src/action/open-edit-visual-model-dialog.ts
@@ -7,6 +7,9 @@ import {
   EditVisualModelDialogState,
 } from "../dialog/visual-model/create-visual-model";
 
+/**
+ * Open dialog to edit visual model information.
+ */
 export function openEditVisualModelDialogAction(
   options: Options,
   dialogs: DialogApiContextType,
