+++ applications/conceptual-model-editor/src/action/shift-attribute.ts
@@ -6,7 +6,11 @@ export enum ShiftAttributeDirection {
   Down = 1,
 };
 
-// TODO RadStr: Maybe should put in addional checks that shiftDistance > 0 and it is integer.
+/**
+ * Moves attribute by {@link shiftDistance} in direction given in {@link shiftDirection}
+ * @param shiftDistance should be an integer.
+ *  Since we modulo, the actual number can be larger than number of attributes.
+ */
 export function shiftAttributePositionAction(
   notifications: UseNotificationServiceWriterType,
   visualModel: WritableVisualModel,
